move stuff around to avoid exposing additional headers
