+++ pvtoolsSrc/nameServerImpl.cpp
@@ -5,7 +5,9 @@
  */
 
 
+#define epicsExportSharedSymbols
 #include <pv/hexDump.h>
+#include <pv/logger.h>
 #include <pv/stringUtility.h>
 
 #include "nameServerImpl.h"
@@ -52,11 +54,11 @@ void NameServerImpl::discoverServers(ServerAddressList& serverAddressList)
         size_t count = entry.addresses.size();
         std::string addresses = " ";
         for (size_t i = 0; i < count; i++) {
-            addresses = addresses + inetAddressToString(entry.addresses[i]) + " ";
+            addresses = addresses + inetAddrToString(entry.addresses[i]) + " ";
         }
         LOG(logLevelDebug, "Found server GUID 0x%s version %d: %s@[%s]", entry.guid.c_str(), (int)entry.version, entry.protocol.c_str(), addresses.c_str());
         if (count > 0) {
-            std::string serverAddress = inetAddressToString(entry.addresses[0]);
+            std::string serverAddress = inetAddrToString(entry.addresses[0]);
             if (addUniqueServerToList(serverAddress, serverAddressList)) {
                 nDiscoveredServers++;
             }
